Refactor shared GPU tile op lowerers
