+++ src/backend/common/op/fill.h
@@ -1,0 +1,82 @@
+/*!
+ * \file tl/backend/common/op/fill.h
+ * \brief Shared tl.fill lowering for GPU backends.
+ */
+
+#ifndef TVM_TL_BACKEND_COMMON_OP_FILL_H_
+#define TVM_TL_BACKEND_COMMON_OP_FILL_H_
+
+#include "op/fill.h"
+
+#include "op/utils.h"
+#include "transform/loop_partition.h"
+#include "transform/loop_vectorize.h"
+
+namespace tvm {
+namespace tl {
+namespace backend {
+
+using namespace tir;
+
+struct Fill {
+  static Stmt Lower(const FillNode &op, const LowerArgs &T,
+                    arith::Analyzer *analyzer) {
+    if (IsFragmentBuffer(op.dst)) {
+      auto par_op = ParallelOp(op.MakeSIMTLoop(analyzer));
+      par_op->InferLayout({T.target,
+                           T.thread_bounds,
+                           T.layout_map,
+                           analyzer,
+                           false,
+                           T.buffer_remap,
+                           {}},
+                          InferLevel::kFree);
+      auto thread_loop = PartitionLoop(par_op->GetRoot(), T.thread_var,
+                                       analyzer, par_op->GetLoopLayout());
+      auto vectorized_loop = VectorizeLoop(thread_loop, analyzer, T.layout_map);
+      auto unrolled_loop = PragmaUnrollLoop(vectorized_loop);
+
+      if (par_op->GetPredicate(T.thread_var).defined()) {
+        return IfThenElse(par_op->GetPredicate(T.thread_var).value(),
+                          unrolled_loop);
+      }
+      return unrolled_loop;
+    }
+
+    if (IsLocalBuffer(op.dst) || IsLocalVarBuffer(op.dst)) {
+      auto init_loop = op.MakeSIMTLoop(analyzer);
+      auto vectorized_loop = VectorizeLoop(init_loop, analyzer, T.layout_map);
+      return PragmaUnrollLoop(vectorized_loop);
+    }
+
+    if (IsSharedBuffer(op.dst) || IsGlobalBuffer(op.dst)) {
+      auto par_op = ParallelOp(op.MakeSIMTLoop(analyzer));
+      par_op->InferLayout({T.target,
+                           T.thread_bounds,
+                           T.layout_map,
+                           analyzer,
+                           false,
+                           T.buffer_remap,
+                           {}},
+                          InferLevel::kFree);
+      auto thread_loop = PartitionLoop(par_op->GetRoot(), T.thread_var,
+                                       analyzer, par_op->GetLoopLayout());
+      auto vectorized_loop = VectorizeLoop(thread_loop, analyzer, T.layout_map);
+      auto unrolled_loop = PragmaUnrollLoop(vectorized_loop);
+      if (par_op->GetPredicate(T.thread_var).defined()) {
+        return IfThenElse(par_op->GetPredicate(T.thread_var).value(),
+                          unrolled_loop);
+      }
+      return unrolled_loop;
+    }
+
+    LOG(FATAL) << "Unsupported scope " << op.dst.scope();
+    return Stmt();
+  }
+};
+
+} // namespace backend
+} // namespace tl
+} // namespace tvm
+
+#endif // TVM_TL_BACKEND_COMMON_OP_FILL_H_
